Online status when in the app (except games ect.... )
